# services/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/goals": {
      "post": {
        "operationId": "GoalsController_create",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "objective": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 5000
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 5000,
                    "nullable": true
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "CHAT",
                      "CONTENT",
                      "CAMPAIGN",
                      "RESEARCH",
                      "AUTOMATION",
                      "PUBLISHING",
                      "MULTI_STEP"
                    ]
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["CRITICAL", "HIGH", "NORMAL", "LOW"]
                  },
                  "constraints": {
                    "type": "object",
                    "properties": {
                      "maxCostUsd": {
                        "type": "number",
                        "exclusiveMinimum": true,
                        "minimum": 0,
                        "maximum": 10000
                      },
                      "timeoutMs": {
                        "type": "integer",
                        "exclusiveMinimum": true,
                        "minimum": 0,
                        "maximum": 86400000
                      },
                      "approval": {
                        "type": "boolean"
                      },
                      "provider": {
                        "type": "string",
                        "maxLength": 60
                      },
                      "language": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "retry": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 10
                      }
                    },
                    "additionalProperties": false
                  },
                  "inputs": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "outputs": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "CONTENT",
                        "IMAGE",
                        "VIDEO",
                        "PDF",
                        "POST",
                        "NOTIFICATION",
                        "ANALYTICS"
                      ]
                    }
                  },
                  "schedule": {
                    "type": "object",
                    "properties": {
                      "cron": {
                        "type": "string",
                        "minLength": 9,
                        "maxLength": 100
                      },
                      "timezone": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 64
                      }
                    },
                    "required": ["cron", "timezone"],
                    "additionalProperties": false,
                    "nullable": true
                  }
                },
                "required": ["title", "objective"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Goals"]
      },
      "get": {
        "operationId": "GoalsController_list",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Goals"]
      }
    },
    "/api/v1/goals/{id}": {
      "delete": {
        "operationId": "GoalsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Goals"]
      },
      "get": {
        "operationId": "GoalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Goals"]
      }
    },
    "/api/v1/goals/{id}/executions": {
      "post": {
        "operationId": "GoalsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": ["Goals"]
      }
    },
...
    "/api/v1/memory": {
      "get": {
        "operationId": "MemoryController_list",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Memory"]
      },
      "put": {
        "operationId": "MemoryController_remember",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "value": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  }
                },
                "required": ["key", "value"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Memory"]
      }
    },
    "/api/v1/memory/{id}": {
      "delete": {
        "operationId": "MemoryController_forget",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Memory"]
      }
    },
...
  },
  "info": {
    "title": "AI Social OS API",
    "description": "Mọi route đều cần đăng nhập trừ khi ghi rõ là công khai. Route thuộc phạm vi workspace cần thêm header `X-Workspace-Id`; thiếu nó là 400, và trỏ vào workspace không phải của mình là 404 chứ không phải 403 — 403 sẽ xác nhận workspace đó có tồn tại.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}